refactor(google): share dictionary markdown formatter
