+++ src/features/translation/providers/GoogleTranslate.test.js
@@ -164,7 +164,7 @@ describe('GoogleTranslateProvider newline chunk isolation', () => {
         'Noun: test, experiment\nVerb: try, attempt'
       );
 
-      expect(result).toBe('Noun: test, experiment\nVerb: try, attempt');
+      expect(result).toBe('**Noun**: test, experiment\n**Verb**: try, attempt');
     });
 
     it('formats dj=1 JSON candidates as the current Markdown contract', async () => {
@@ -176,7 +176,7 @@ describe('GoogleTranslateProvider newline chunk isolation', () => {
       });
 
       expect(result).toBe(
-        'Noun: test, experiment\nPronunciation: /tɛst/\n\nDefinitions:\n- (noun) a test\n\nExamples:\n- This is a test'
+        '**Noun**: test, experiment\n\n**Pronunciation**: /tɛst/\n\n**Definitions**:\n- (noun) a test\n\n**Examples**:\n- This is a test'
       );
     });
 
@@ -200,7 +200,7 @@ describe('GoogleTranslateProvider newline chunk isolation', () => {
 
       expect(classicResult).toBe(v2Result);
       expect(classicResult).toBe(
-        'Noun: test, experiment\nPronunciation: /tɛst/\n\nDefinitions:\n- (noun) a test\n\nExamples:\n- This is a test'
+        '**Noun**: test, experiment\n\n**Pronunciation**: /tɛst/\n\n**Definitions**:\n- (noun) a test\n\n**Examples**:\n- This is a test'
       );
     });
   });
@@ -262,7 +262,7 @@ describe('GoogleTranslateProvider newline chunk isolation', () => {
       );
 
       expect(result).toBe(
-        'translated text\n\nNoun: test, experiment\nPronunciation: /tɛst/\n\nDefinitions:\n- (noun) a test\n\nExamples:\n- This is a test'
+        'translated text\n\n**Noun**: test, experiment\n\n**Pronunciation**: /tɛst/\n\n**Definitions**:\n- (noun) a test\n\n**Examples**:\n- This is a test'
       );
     });
   });
